first day of code
This is the first day of coding, where we simply added an avatar icon on the left side of the notification and cart . This is only for UI. We also added a search bar field where we can enter anything according to our requirements, but the search function is not working right now. Then, we have a products tab bar where we can select products. Currently, only the product API is working; however, we plan to modify it tomorrow to include men's and women's product APIs. The product API is displayed in the container that we designed. Lastly, there is a bottom navigator bar with home, notification, cart, and profile pages.